refactor(storybook): story location & issues
